# docs/openapi.legacy-monolith.yaml
openapi: 3.0.3
info:
  title: Splitward
  version: 0.1.0
  description: Multi-hospital split-learning control plane that trains shared nets without raw patient
    exchange.
  x-product-slug: ai-decentralized-splitward
  x-source: ai-in-decentralized+ai/raskarsplitlearningmit2018-181129165110/
servers:
- url: https://api.example.com
...
paths:
...
  /v1/studies/{studyId}/approve-privacy:
    parameters:
    - $ref: '#/components/parameters/StudyId'
    post:
      operationId: approveStudyPrivacy
      tags:
      - Studies
      summary: Privacy officer approval
      security:
      - BearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataEnvelopeStudy'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /v1/contribution-credits:
    get:
      operationId: listContributionCredits
      tags:
      - ContributionCredits
      summary: List ContributionCredit
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      parameters:
      - $ref: '#/components/parameters/Cursor'
      - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListEnvelopeContributionCredit'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      operationId: recordContributionCredit
      tags:
      - ContributionCredits
      summary: Create ContributionCredit
      security:
      - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContributionCreditCreate'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataEnvelopeContributionCredit'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/contribution-credits/{creditId}:
    parameters:
    - $ref: '#/components/parameters/CreditId'
    get:
      operationId: getContributionCredit
      tags:
      - ContributionCredits
      summary: Get ContributionCredit
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataEnvelopeContributionCredit'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /v1/incidents/{incidentId}/halt-study:
    parameters:
    - $ref: '#/components/parameters/IncidentId'
    post:
      operationId: haltStudyForIncident
      tags:
      - Incidents
      summary: Halt study on incident
      security:
      - BearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataEnvelopeIncident'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
...
    StudyId:
      name: studyId
      in: path
      required: true
      schema:
        type: string
...
    CreditId:
      name: creditId
      in: path
      required: true
      schema:
        type: string
    IncidentId:
      name: incidentId
      in: path
      required: true
      schema:
        type: string
...
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    BadRequest:
      description: Bad request
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: Not found
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Conflict:
      description: Conflict
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
    ResponseMeta:
      type: object
      required:
      - requestId
      - timestamp
      properties:
        requestId:
          type: string
        timestamp:
          type: string
          format: date-time
    Problem:
      type: object
      required:
      - type
      - title
      - status
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
...
    Study:
      type: object
      required:
      - id
      - title
      - strategy
      - status
      - privacyApprovalStatus
      properties:
        id:
          type: string
        title:
          type: string
        strategy:
          type: string
          enum:
          - split
          - federated
          - hybrid
        status:
          type: string
          enum:
          - draft
          - approved
          - running
          - halted
          - completed
        privacyApprovalStatus:
          type: string
          enum:
          - pending
          - approved
          - rejected
        activationTtlHours:
          type: integer
...
    DataEnvelopeStudy:
      type: object
      required:
      - data
      - meta
      properties:
        data:
          $ref: '#/components/schemas/Study'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
...
    ContributionCredit:
      type: object
      required:
      - id
      - studyId
      - siteId
      - roundsCompleted
      - dataVolumeClass
      properties:
        id:
          type: string
        studyId:
          type: string
        siteId:
          type: string
        roundsCompleted:
          type: integer
        dataVolumeClass:
          type: string
          enum:
          - S
          - M
          - L
          - XL
        creditPoints:
          type: number
    ContributionCreditCreate:
      type: object
      required:
      - studyId
      - siteId
      - roundsCompleted
      - dataVolumeClass
      properties:
        studyId:
          type: string
        siteId:
          type: string
        roundsCompleted:
          type: integer
        dataVolumeClass:
          type: string
          enum:
          - S
          - M
          - L
          - XL
        creditPoints:
          type: number
    DataEnvelopeContributionCredit:
      type: object
      required:
      - data
      - meta
      properties:
        data:
          $ref: '#/components/schemas/ContributionCredit'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    ListEnvelopeContributionCredit:
      type: object
      required:
      - data
      - meta
      properties:
        data:
          type: object
          required:
          - items
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/ContributionCredit'
            nextCursor:
              type: string
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    Incident:
      type: object
      required:
      - id
      - studyId
      - incidentType
      - severity
      - status
      properties:
        id:
          type: string
        studyId:
          type: string
        incidentType:
          type: string
          enum:
          - suspectedLeakage
          - protocolViolation
          - egressBreach
        severity:
          type: string
          enum:
          - low
          - medium
          - high
          - critical
        status:
          type: string
          enum:
          - open
          - quarantined
          - closed
        detail:
          type: string
...
    DataEnvelopeIncident:
      type: object
      required:
      - data
      - meta
      properties:
        data:
          $ref: '#/components/schemas/Incident'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
...