# platform/api-server/src/domains/contribution-credits/handlers/contribution-credit.handlers.ts
/**
 * Handlers for ContributionCredit
 *
 * Generated by Handler Generator v2.0.0
 * Generator Version: 2.0.0
 * Domain: contribution-credits
 *
 * ⚠️  DO NOT EDIT THIS FILE MANUALLY
 * This file is auto-generated. Any manual changes will be overwritten.
 */

import type { FastifyRequest, FastifyReply } from "fastify";
import type { ContributionCreditsDomainModule } from "../dependencies/contribution-credits-ddd.dependencies.js";
import type { GetContributionCreditInput, ListContributionCreditsInput, RecordContributionCreditInput } from "@splitward/services/contribution-credits";
// Use cases are available via deps.useCases.<repoKey>.<verb>
// Execution context is automatically set up by route wrapper (wrapHandlerInExecutionContext)

/**
 * Handler for GET /v0/tenants/me/contribution-credits
 * Operation: listContributionCredits
 */
export async function listContributionCredits(
  request: FastifyRequest<{ Params: Record<string, string>; Querystring: { cursor?: string | string[], limit?: string | string[] }; Body: unknown }>,
  reply: FastifyReply,
  deps: ContributionCreditsDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: ListContributionCreditsInput = {
    ...(request.query ?? {}),
  };
  const result = await deps.useCases.contributionCredits.list.execute(input);
  return reply.code(200).send(result);
}


/**
 * Handler for POST /v0/tenants/me/contribution-credits
 * Operation: recordContributionCredit
 */
export async function recordContributionCredit(
  request: FastifyRequest,
  reply: FastifyReply,
  deps: ContributionCreditsDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: RecordContributionCreditInput = {
    ...(request.body ?? {}),
  };
  const result = await deps.useCases.contributionCredits.get.execute(input);
  return reply.code(201).send(result);
}


/**
 * Handler for GET /v0/tenants/me/contribution-credits/{creditId}
 * Operation: getContributionCredit
 */
export async function getContributionCredit(
  request: FastifyRequest<{ Params: { creditId: string }; Querystring: Record<string, string | string[] | undefined>; Body: unknown }>,
  reply: FastifyReply,
  deps: ContributionCreditsDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: GetContributionCreditInput = {
    creditId: request.params.creditId,
  };
  const result = await deps.useCases.contributionCredits.get.execute(input);
  return reply.code(200).send(result);
}