# platform/api-server/src/domains/incidents/handlers/halt-study.handlers.ts
/**
 * Handlers for HaltStudy
 *
 * Generated by Handler Generator v2.0.0
 * Generator Version: 2.0.0
 * Domain: incidents
 *
 * ⚠️  DO NOT EDIT THIS FILE MANUALLY
 * This file is auto-generated. Any manual changes will be overwritten.
 */

import type { FastifyRequest, FastifyReply } from "fastify";
import type { IncidentsDomainModule } from "../dependencies/incidents-ddd.dependencies.js";
import type { HaltStudyForIncidentInput } from "@splitward/services/incidents";
// Use cases are available via deps.useCases.<repoKey>.<verb>
// Execution context is automatically set up by route wrapper (wrapHandlerInExecutionContext)

/**
 * Handler for POST /v0/tenants/me/incidents/{incidentId}/halt-study
 * Operation: haltStudyForIncident
 */
export async function haltStudyForIncident(
  request: FastifyRequest<{ Params: { incidentId: string }; Querystring: Record<string, string | string[] | undefined>; Body: unknown }>,
  reply: FastifyReply,
  deps: IncidentsDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: HaltStudyForIncidentInput = {
    incidentId: request.params.incidentId,
    ...(request.body ?? {}),
  };
  const result = await deps.useCases.haltStudies.get.execute(input);
  return reply.code(200).send(result);
}