# platform/api-server/src/domains/studies/handlers/approve-privacy.handlers.ts
/**
 * Handlers for ApprovePrivacy
 *
 * Generated by Handler Generator v2.0.0
 * Generator Version: 2.0.0
 * Domain: studies
 *
 * ⚠️  DO NOT EDIT THIS FILE MANUALLY
 * This file is auto-generated. Any manual changes will be overwritten.
 */

import type { FastifyRequest, FastifyReply } from "fastify";
import type { StudiesDomainModule } from "../dependencies/studies-ddd.dependencies.js";
import type { ApproveStudyPrivacyInput } from "@splitward/services/studies";
// Use cases are available via deps.useCases.<repoKey>.<verb>
// Execution context is automatically set up by route wrapper (wrapHandlerInExecutionContext)

/**
 * Handler for POST /v0/tenants/me/studies/{studyId}/approve-privacy
 * Operation: approveStudyPrivacy
 */
export async function approveStudyPrivacy(
  request: FastifyRequest<{ Params: { studyId: string }; Querystring: Record<string, string | string[] | undefined>; Body: unknown }>,
  reply: FastifyReply,
  deps: StudiesDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: ApproveStudyPrivacyInput = {
    studyId: request.params.studyId,
    ...(request.body ?? {}),
  };
  const result = await deps.useCases.approvePrivacies.create.execute(input);
  return reply.code(200).send(result);
}